refactor(CQ2-3): split sync_service.py (1181 lines) into sync_service/ package
- Split SyncService into 4 files using mixin pattern:
  - __init__.py: SyncService facade + singleton-dependent methods
  - task_manager.py: TaskManagerMixin (task lifecycle + media server delegation)
  - retry.py: RetryMixin (_retry_mark_episode sync/async)
  - season_info.py: SeasonInfoMixin (season keyword detection)
- Re-export asyncio/time for test patch compatibility
- Convert top-level `from ..sync_service import sync_service` to late imports
  in 6 subpackage files (emby/jellyfin/plex/custom/feiniu/fongmi) to break
  circular import introduced when sync_service became a package
- All 1894 tests pass; ruff clean
