+++ app/services/emby/sync_service.py
@@ -11,7 +11,6 @@ from typing import Any
 
 from ...core.logging import logger
 from ...models.sync import SyncResponse
-from ..sync_service import sync_service
 from .extractor import extract_emby_data
 
 EMBY_SYNC_SOURCE = "emby"
@@ -32,6 +31,8 @@ class EmbySyncService:
             sync_svc: 共享 SyncService 实例。未传入时使用模块级单例。
         """
         if sync_svc is None:
+            from ..sync_service import sync_service
+
             sync_svc = sync_service
         try:
             # 记录接收到的数据
